+++ src/locale/Locale.ts
@@ -37,6 +37,7 @@ export const enUS = {
 		EditVenueEditorAmount: 'Unable to edit the editor compensation',
 		EditVenueSubmissionCost: 'Unable to edit the submission cost',
 		EditRoleBidding: 'Unable to toggle bidding',
+		EditRoleDesiredAssignments: 'Unable to set desired assignments',
 		EditRoleApprover: 'Unable to set approver',
 		CreateRole: 'Unable to create new role.',
 		UpdateRoleName: 'Unable to update role name',